requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
pytz==2023.3
gunicorn==21.2.0